update destroy layanan

+++ app/Http/Controllers/LayananController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Layanan;
+use App\Models\LayananKonter;
 use Illuminate\Http\Request;
 
 class LayananController extends Controller
@@ -48,7 +49,12 @@ class LayananController extends Controller
     public function destroy($id)
     {
         $layanan = Layanan::find($id);
-        $layanan->delete();
-        return redirect()->back()->with('success', 'Berhasil menghapus data');
+        $cek_layanan_konter = LayananKonter::where('id_layanan', $id)->count();
+        if ($cek_layanan_konter != 0) {
+            return redirect()->back()->with('danger', 'Layanan telah digunakan oleh konter');
+        } else {
+            $layanan->delete();
+            return redirect()->back()->with('success', 'Berhasil menghapus data');
+        }
     }
 }
